VozPyt/VozPyt/requirements.txt
SpeechRecognition
PyAudio
psycopg2-binary
python-dotenv